requirements.txt
streamlit
numpy
pandas
scipy
statsmodels
joblib
mysql-connector-python
nltk
mysql.connector
matplotlib
tensorflow
scikit learn